+++ internal/app/wwctl/kernel/root.go
@@ -2,8 +2,6 @@ package kernel
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/warewulf/warewulf/internal/app/wwctl/kernel/delete"
-	"github.com/warewulf/warewulf/internal/app/wwctl/kernel/imprt"
 	"github.com/warewulf/warewulf/internal/app/wwctl/kernel/list"
 )
 
@@ -11,15 +9,13 @@ var (
 	baseCmd = &cobra.Command{
 		DisableFlagsInUseLine: true,
 		Use:                   "kernel COMMAND [OPTIONS]",
-		Short:                 "Kernel Image Management",
-		Long:                  "This command manages Warewulf Kernels used for bootstrapping nodes",
+		Short:                 "Kernel Management",
+		Long:                  "This command manages kernels available to Warewulf",
 	}
 )
 
 func init() {
-	baseCmd.AddCommand(imprt.GetCommand())
 	baseCmd.AddCommand(list.GetCommand())
-	baseCmd.AddCommand(delete.GetCommand())
 }
 
 // GetRootCommand returns the root cobra.Command for the application.